refactor: rename lir module to tir and update references accordingly
