+++ .github/workflows/ci.yml
@@ -17,7 +17,8 @@ on:
 
 env:
   LLVM_BRANCH: release/20.x
-  LLVM_INSTALL_PATH: ${{ github.workspace }}/llvm-install
+  LLVM_INSTALL_PATH_UNIX: ${{ github.workspace }}/llvm-install
+  LLVM_INSTALL_PATH_WINDOWS: ${{ github.workspace }}\llvm-install
   RUST_BACKTRACE: 1
 
 
@@ -50,24 +51,45 @@ jobs:
       # -------------------------------
       # Restore LLVM cache
       # -------------------------------
-      - name: Restore LLVM cache
-        id: cache-llvm-restore
+      - name: Restore LLVM cache (Linux/macOS)
+        id: cache-llvm-restore-unix
         uses: actions/cache/restore@v4
         with:
-          path: ${{ env.LLVM_INSTALL_PATH }}
+          path: ${{ env.LLVM_INSTALL_PATH_UNIX }}
+          key: ${{ runner.os }}-${{ runner.arch }}-llvm-${{ env.LLVM_BRANCH }}
+
+      - name: Restore LLVM cache (Windows)
+        id: cache-llvm-restore-windows
+        if: runner.os == 'Windows'
+        uses: actions/cache/restore@v4
+        with:
+          path: ${{ env.LLVM_INSTALL_PATH_WINDOWS }}
           key: ${{ runner.os }}-${{ runner.arch }}-llvm-${{ env.LLVM_BRANCH }}
 
       # -------------------------------
       # Build LLVM if cache missed
       # -------------------------------
-      - name: Build LLVM
+      - name: Build LLVM (Linux/macOS)
         if: steps.cache-llvm-restore.outputs.cache-hit != 'true'
         shell: bash
         run: |
           git clone --depth 1 --branch ${{ env.LLVM_BRANCH }} https://github.com/llvm/llvm-project.git
           cd llvm-project
           cmake -S llvm -B build -G Ninja \
-            -DCMAKE_INSTALL_PREFIX=${{ env.LLVM_INSTALL_PATH }} \
+            -DCMAKE_INSTALL_PREFIX=${{ env.LLVM_INSTALL_PATH_UNIX }} \
+            -DCMAKE_BUILD_TYPE=Release \
+            -DLLVM_ENABLE_PROJECTS="clang;lld" \
+            -DLLVM_TARGETS_TO_BUILD="X86;AArch64"
+          ninja -C build install
+
+      - name: Build LLVM (Windows)
+        if: steps.cache-llvm-restore.outputs.cache-hit != 'true' && runner.os == 'Windows'
+        shell: bash
+        run: |
+          git clone --depth 1 --branch ${{ env.LLVM_BRANCH }} https://github.com/llvm/llvm-project.git
+          cd llvm-project
+          cmake -S llvm -B build -G Ninja \
+            -DCMAKE_INSTALL_PREFIX=${{ env.LLVM_INSTALL_PATH_WINDOWS }} \
             -DCMAKE_BUILD_TYPE=Release \
             -DLLVM_ENABLE_PROJECTS="clang;lld" \
             -DLLVM_TARGETS_TO_BUILD="X86;AArch64"
@@ -78,22 +100,29 @@ jobs:
       # -------------------------------
       - name: Ensure LLVM cache path exists
         if: always()
-        run: mkdir -p ${{ env.LLVM_INSTALL_PATH }}
+        run: mkdir -p ${{ env.LLVM_INSTALL_PATH_UNIX }}
         shell: bash
 
       - name: Ensure LLVM cache path exists (Windows)
         if: always() && runner.os == 'Windows'
         shell: pwsh
-        run: New-Item -ItemType Directory -Force -Path $env:LLVM_INSTALL_PATH
+        run: New-Item -ItemType Directory -Force -Path $env:LLVM_INSTALL_PATH_WINDOWS
 
       # -------------------------------
       # Save LLVM cache always
       # -------------------------------
-      - name: Save LLVM cache
+      - name: Save LLVM cache (Linux/macOS)
         if: always()
         uses: actions/cache/save@v4
         with:
-          path: ${{ env.LLVM_INSTALL_PATH }}
+          path: ${{ env.LLVM_INSTALL_PATH_UNIX }}
+          key: ${{ runner.os }}-${{ runner.arch }}-llvm-${{ env.LLVM_BRANCH }}
+
+      - name: Save LLVM cache (Windows)
+        if: always() && runner.os == 'Windows'
+        uses: actions/cache/save@v4
+        with:
+          path: ${{ env.LLVM_INSTALL_PATH_UNIX }}
           key: ${{ runner.os }}-${{ runner.arch }}-llvm-${{ env.LLVM_BRANCH }}
       
       # -------------------------------
@@ -105,23 +134,33 @@ jobs:
           toolchain: stable
           profile: minimal
 
-      # -------------------------------
-      # Add LLVM to PATH
-      # -------------------------------
-      - name: Add LLVM to path (Unix)
-        if: runner.os != 'Windows'
-        run: echo "${{ env.LLVM_INSTALL_PATH }}/bin" >> $GITHUB_PATH
 
+      # -------------------------------
+      # Set LLVM_SYS_201_PREFIX environment variable for llvm-sys crate
+      # -------------------------------
+      - name: Set LLVM_SYS_201_PREFIX (Unix)
+        shell: bash
+        run: |
+          echo "LLVM_SYS_201_PREFIX=${{ env.LLVM_INSTALL_PATH_UNIX }}" >> $GITHUB_ENV
+
+      
       - name: Set LLVM_SYS_201_PREFIX (Windows)
         if: runner.os == 'Windows'
         shell: pwsh
         run: |
-          echo "LLVM_SYS_201_PREFIX=${{ env.LLVM_INSTALL_PATH }}" | Out-File -FilePath $env:GITHUB_ENV -Encoding utf8 -Append
+          echo "LLVM_SYS_201_PREFIX=${{ env.LLVM_INSTALL_PATH_WINDOWS }}" | Out-File -FilePath $env:GITHUB_ENV -Encoding utf8 -Append
+
+      # -------------------------------
+      # Add LLVM to PATH
+      # -------------------------------
+      - name: Add LLVM to path (Unix)
+        shell: bash
+        run: echo "${{ env.LLVM_INSTALL_PATH_UNIX }}/bin" >> $GITHUB_PATH
 
       - name: Add LLVM to path (Windows)
         if: runner.os == 'Windows'
         shell: pwsh
-        run: echo "${{ env.LLVM_INSTALL_PATH }}\bin" | Out-File -FilePath $env:GITHUB_PATH -Encoding utf8 -Append
+        run: echo "${{ env.LLVM_INSTALL_PATH_WINDOWS }}\bin" | Out-File -FilePath $env:GITHUB_PATH -Encoding utf8 -Append
 
       # -------------------------------
       # Build and test Rust project
